Centralize icon url build logic
