+++ mobile/src/main/java/org/openhab/habdroid/model/LinkedPage.kt
@@ -18,6 +18,7 @@ import android.os.Parcelable
 import kotlinx.android.parcel.Parcelize
 import org.json.JSONObject
 import org.openhab.habdroid.util.IconFormat
+import org.openhab.habdroid.util.addIconUrlParameters
 import org.openhab.habdroid.util.forEach
 import org.openhab.habdroid.util.optStringOrNull
 import org.w3c.dom.Node
@@ -73,15 +74,11 @@ fun JSONObject?.toLinkedPage(iconFormat: IconFormat): LinkedPage? {
     if (this == null) {
         return null
     }
-    val suffix = when (iconFormat) {
-        IconFormat.Png -> "png"
-        IconFormat.Svg -> "svg"
-    }
     val icon = optStringOrNull("icon")
     return LinkedPage.build(
         getString("id"),
         optStringOrNull("title"),
         icon,
-        "icon/$icon?&format=$suffix&anyFormat=true",
+        "icon/$icon".addIconUrlParameters(iconFormat),
         getString("link"))
 }
